feat: added theme support for icons - UiThemeProvider / revert class
