+++ apps/design-system/src/App.tsx
@@ -4,6 +4,9 @@ import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom'
 import ComponentPage from '@/pages/component-page'
 import ViewPreview from '@/pages/view-preview/view-preview'
 import DocsLayout from '@components/docs-layout/docs-layout'
+import { useThemeStore } from '@utils/theme-utils'
+
+import { ThemeProvider } from '@harnessio/ui/context'
 
 const router = createBrowserRouter([
   { path: '/view-preview/*', element: <ViewPreview /> },
@@ -12,7 +15,13 @@ const router = createBrowserRouter([
 ])
 
 const App: FC = () => {
-  return <RouterProvider router={router} />
+  const themeStore = useThemeStore()
+
+  return (
+    <ThemeProvider {...themeStore}>
+      <RouterProvider router={router} />
+    </ThemeProvider>
+  )
 }
 
 export default App
