+++ apps/design-system/src/subjects/views/repo-empty/repo-empty-view.tsx
@@ -1,16 +1,45 @@
-import { noop } from '@utils/viewUtils'
+import { FC, useMemo } from 'react'
 
-import { RepoEmptyView } from '@harnessio/ui/views'
+import repoSummaryProps from '@subjects/views/repo-summary/repo-summary-props.json'
+import { noop, useTranslationStore } from '@utils/viewUtils'
 
-export const RepoEmpty = () => {
+import { BranchSelectorV2, RepoSummaryView, RepoSummaryViewProps } from '@harnessio/ui/views'
+
+export const RepoEmpty: FC<Partial<RepoSummaryViewProps>> = props => {
+  const repoEntryPathToFileTypeMap = useMemo<RepoSummaryViewProps['repoEntryPathToFileTypeMap']>(() => {
+    return new Map()
+  }, [])
+
   return (
-    <RepoEmptyView
-      httpUrl="https://github.com/mock-repo"
-      repoName="mock-repo"
-      projName="mock-project"
-      sshUrl="git@github.com:mock-repo.git"
-      gitRef=""
+    <RepoSummaryView
+      {...repoSummaryProps}
+      repoEntryPathToFileTypeMap={repoEntryPathToFileTypeMap}
+      selectedBranchOrTag={{ name: 'main', sha: '' }}
+      saveDescription={noop}
       handleCreateToken={noop}
+      navigateToFile={noop}
+      useTranslationStore={useTranslationStore}
+      gitRef=""
+      updateRepoError=""
+      isEditDialogOpen={false}
+      setEditDialogOpen={noop}
+      searchQuery=""
+      setSearchQuery={noop}
+      branchSelectorRenderer={
+        <BranchSelectorV2
+          repoId="canary"
+          spaceId="org"
+          branchList={[]}
+          tagList={[]}
+          selectedBranchorTag={{ name: 'main', sha: 'sha' }}
+          onSelectBranch={noop}
+          isBranchOnly={false}
+          dynamicWidth={false}
+          useTranslationStore={useTranslationStore}
+          setSearchQuery={noop}
+        />
+      }
+      {...props}
     />
   )
 }
