+++ apps/design-system/src/utils/viewUtils.ts
@@ -15,5 +15,3 @@ export const mockT = (key: string, options?: { [key: string]: any }) => {
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const useTranslationStore = () => ({ t: mockT as any, changeLanguage: noop, i18n: {} as any })
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const useThemeStore = () => ({ theme: 'dark-std-std' as any, setTheme: noop })
