+++ packages/ui/src/components/icon/icon-name-map.ts
@@ -1,0 +1,440 @@
+import * as React from 'react'
+
+import AccountIcon from '../../icons/account-icon.svg'
+import Account from '../../icons/account.svg'
+import AddFile from '../../icons/add-file-icon.svg'
+import AddFolder from '../../icons/add-folder-icon.svg'
+import AISparks from '../../icons/ai-sparks.svg'
+import AppleShortcut from '../../icons/apple-shortcut.svg'
+import ArrowLong from '../../icons/arrow-long.svg'
+import ArrowShort from '../../icons/arrow-short.svg'
+import Arrow from '../../icons/arrow.svg'
+import ArtifactsGradient from '../../icons/artifacts-gradient.svg'
+import Artifacts from '../../icons/artifacts-icon.svg'
+import AttachmentImage from '../../icons/attachment-image.svg'
+import Attachment from '../../icons/attachment.svg'
+import BitrisePlugin from '../../icons/bitrise-plugin.svg'
+import BoldPlus from '../../icons/bold-plus.svg'
+import Bold from '../../icons/bold.svg'
+import BookmarkAdd from '../../icons/bookmark-add.svg'
+import BookmarkIcon from '../../icons/bookmark-icon.svg'
+import BoxCloning from '../../icons/box-cloning.svg'
+import BoxGuide from '../../icons/box-guide.svg'
+import BoxLightning from '../../icons/box-lightning.svg'
+import BoxPullRequests from '../../icons/box-pull-requests.svg'
+import Branch from '../../icons/branch.svg'
+import Briefcase from '../../icons/briefcase-icon.svg'
+import CancelGrey from '../../icons/cancel-grey.svg'
+import Changes from '../../icons/changes.svg'
+import ChaosEngineeringGradient from '../../icons/chaos-engineering-gradient.svg'
+import ChaosEngineering from '../../icons/chaos-engineering-icon.svg'
+import CheckboxCircle from '../../icons/checkbox-circle.svg'
+import Checkbox from '../../icons/checkbox.svg'
+import Checklist from '../../icons/checklist.svg'
+import Checks from '../../icons/checks.svg'
+import ChevronDown from '../../icons/chevron-down.svg'
+import ChevronFillDown from '../../icons/chevron-fill-down.svg'
+import ChevronRight from '../../icons/chevron-right.svg'
+import ChevronUp from '../../icons/chevron-up.svg'
+import CircleArrowTopRight from '../../icons/circle-arrow-top-right.svg'
+import CircleArrowTop from '../../icons/circle-arrow-top.svg'
+import CircleArrowsUpDown from '../../icons/circle-arrows-updown.svg'
+import CirclePlus from '../../icons/circle-plus.svg'
+import CircleWithSector from '../../icons/circle-with-sector.svg'
+import Circle from '../../icons/circle.svg'
+import ClockIcon from '../../icons/clock-icon.svg'
+import Clock from '../../icons/clock.svg'
+import Clone from '../../icons/clone.svg'
+import Close from '../../icons/close.svg'
+import CloudCostsGradient from '../../icons/cloud-costs-gradient.svg'
+import CloudCosts from '../../icons/cloud-costs-icon.svg'
+import CodeBrackets from '../../icons/code-brackets.svg'
+import Code from '../../icons/code.svg'
+import Cog6 from '../../icons/cog-6.svg'
+import CollapseComment from '../../icons/collapse-comment.svg'
+import CollapseDiff from '../../icons/collapse-diff.svg'
+import CollapseIn from '../../icons/collapse-in.svg'
+import CollapseOut from '../../icons/collapse-out.svg'
+import CommandSymbol from '../../icons/command-symbol.svg'
+import Comments from '../../icons/comments.svg'
+import Compare from '../../icons/compare.svg'
+import Connectors from '../../icons/connectors-icon.svg'
+import CrossCircle from '../../icons/cross-circle.svg'
+import Cross from '../../icons/cross.svg'
+import DashboardsGradient from '../../icons/dashboards-gradient.svg'
+import Dashboards from '../../icons/dashboards-icon.svg'
+import DatabaseGradient from '../../icons/database-gradient.svg'
+import Database from '../../icons/database-icon.svg'
+import DevEnvsGradient from '../../icons/dev-envs-gradient.svg'
+import DevEnvs from '../../icons/dev-envs-icon.svg'
+import DevInsightsGradient from '../../icons/dev-insights-gradient.svg'
+import DevInsights from '../../icons/dev-insights-icon.svg'
+import DevPortalGradient from '../../icons/dev-portal-gradient.svg'
+import DevPortal from '../../icons/dev-portal-icon.svg'
+import Docker from '../../icons/docker.svg'
+import MoreDotsFill from '../../icons/dots-icon.svg'
+import DoubleTick from '../../icons/double-tick.svg'
+import Download from '../../icons/download-icon.svg'
+import Edit from '../../icons/edit-icon.svg'
+import Environment from '../../icons/environment-icon.svg'
+import ExecutionGradient from '../../icons/execution-gradient.svg'
+import Execution from '../../icons/execution-icon.svg'
+import ExpandComment from '../../icons/expand-comment.svg'
+import ExpandDiff from '../../icons/expand-diff.svg'
+import Eye from '../../icons/eye-icon.svg'
+import Fail from '../../icons/fail.svg'
+import FileIcon from '../../icons/file-icon.svg'
+import File from '../../icons/file.svg'
+import FilterList from '../../icons/filter-list.svg'
+import FlagGradient from '../../icons/flag-gradient.svg'
+import Flag from '../../icons/flag-icon.svg'
+import FolderIcon from '../../icons/folder-icon.svg'
+import Folder from '../../icons/folder.svg'
+import GitBranch from '../../icons/git-branch.svg'
+import GithubActions from '../../icons/github-actions.svg'
+import GitnessLogo from '../../icons/gitness-logo.svg'
+import GreenCircle from '../../icons/green-circle.svg'
+import GridDots from '../../icons/grid-dots.svg'
+import HarnessLogoText from '../../icons/harness-logo-text.svg'
+import HarnessPlugin from '../../icons/harness-plugin.svg'
+import Harness from '../../icons/harness.svg'
+import Header from '../../icons/header.svg'
+import Hierarchy from '../../icons/hierarchy-icon.svg'
+import IncidentsGradient from '../../icons/incidents-gradient.svg'
+import Incidents from '../../icons/incidents-icon.svg'
+import InfoCircle from '../../icons/info-circle.svg'
+import InfrastructureGradient from '../../icons/infrastructure-gradient.svg'
+import Infrastructure from '../../icons/infrastructure-icon.svg'
+import Italicize from '../../icons/italicize.svg'
+import Key from '../../icons/key-icon.svg'
+import Lightning from '../../icons/lightning.svg'
+import List from '../../icons/list.svg'
+// All icons with colors inside the SVG should also be available in the light theme with the {`light`} postfix.
+import CreateWorkspaceLight from '../../icons/lists-data-icons/create-workspace-light.svg'
+import CreateWorkspace from '../../icons/lists-data-icons/create-workspace.svg'
+import NoDataBranchesLight from '../../icons/lists-data-icons/no-data-branches-light.svg'
+import NoDataBranches from '../../icons/lists-data-icons/no-data-branches.svg'
+import NoDataCogLight from '../../icons/lists-data-icons/no-data-cog-light.svg'
+import NoDataCog from '../../icons/lists-data-icons/no-data-cog.svg'
+import NoDataCommitsLight from '../../icons/lists-data-icons/no-data-commits-light.svg'
+import NoDataCommits from '../../icons/lists-data-icons/no-data-commits.svg'
+import NoDataErrorLight from '../../icons/lists-data-icons/no-data-error-light.svg'
+import NoDataError from '../../icons/lists-data-icons/no-data-error.svg'
+import NoDataFolderLight from '../../icons/lists-data-icons/no-data-folder-light.svg'
+import NoDataFolder from '../../icons/lists-data-icons/no-data-folder.svg'
+import NoDataMembersLight from '../../icons/lists-data-icons/no-data-members-light.svg'
+import NoDataMembers from '../../icons/lists-data-icons/no-data-members.svg'
+import NoDataMergeLight from '../../icons/lists-data-icons/no-data-merge-light.svg'
+import NoDataMerge from '../../icons/lists-data-icons/no-data-merge.svg'
+import NoDataPRLight from '../../icons/lists-data-icons/no-data-pr-light.svg'
+import NoDataPR from '../../icons/lists-data-icons/no-data-pr.svg'
+import NoRepositoryLight from '../../icons/lists-data-icons/no-data-repository-light.svg'
+import NoRepository from '../../icons/lists-data-icons/no-data-repository.svg'
+import noDataTags from '../../icons/lists-data-icons/no-data-tags.svg'
+import NoDataWebhooksLight from '../../icons/lists-data-icons/no-data-webhooks-light.svg'
+import NoDataWebhooks from '../../icons/lists-data-icons/no-data-webhooks.svg'
+import NoSearchMagnifyingGlassLight from '../../icons/lists-data-icons/no-search-magnifying-glass-light.svg'
+import NoSearchMagnifyingGlass from '../../icons/lists-data-icons/no-search-magnifying-glass.svg'
+import Lock from '../../icons/lock.svg'
+import LogOut from '../../icons/log-out-icon.svg'
+import HarnessLogoGradientEllipse from '../../icons/logo-gradient-ellipse.svg'
+import HarnessLogoGradient from '../../icons/logo-gradient.svg'
+import MenuDots from '../../icons/menu-dots-icon.svg'
+import Merged from '../../icons/merged.svg'
+import Minus from '../../icons/minus-icon.svg'
+import Ellipsis from '../../icons/more-dots-icon.svg'
+import Navigation from '../../icons/navigation-icon.svg'
+import NodeLogo from '../../icons/node-logo.svg'
+import Notification from '../../icons/notification-icon.svg'
+import OpenPR from '../../icons/open-pr.svg'
+import Paint from '../../icons/paint-icon.svg'
+import PendingClock from '../../icons/pending-clock.svg'
+import Pin from '../../icons/pin.svg'
+import PipelinesGradient from '../../icons/pipelines-gradient.svg'
+import Pipelines from '../../icons/pipelines-icon.svg'
+import Play from '../../icons/play-solid.svg'
+import Plug from '../../icons/plug.svg'
+import Plus from '../../icons/plus.svg'
+import PrClosed from '../../icons/pr-closed.svg'
+import PrComment from '../../icons/pr-comment.svg'
+import PrMerge from '../../icons/pr-merge.svg'
+import PRMerged from '../../icons/pr-merged.svg'
+import PrOpen from '../../icons/pr-open.svg'
+import PrReview from '../../icons/pr-review.svg'
+import Pull from '../../icons/pull-icon.svg'
+import PythonAndNodeLogo from '../../icons/python-and-node-logo.svg'
+import PythonLogo from '../../icons/python-logo.svg'
+import RepoIcon from '../../icons/repo-icon.svg'
+import RepositoriesGradient from '../../icons/repositories-gradient.svg'
+import Repositories from '../../icons/repositories-icon.svg'
+import Rocket from '../../icons/rocket.svg'
+import RunTest from '../../icons/run-test.svg'
+import Run from '../../icons/run.svg'
+import Running from '../../icons/running.svg'
+import Search from '../../icons/search-icon.svg'
+import Secrets from '../../icons/secrets-icon.svg'
+import SecurityTestsGradient from '../../icons/security-tests-gradient.svg'
+import SecurityTests from '../../icons/security-tests-icon.svg'
+import Settings1 from '../../icons/setting-1.svg'
+import Settings2 from '../../icons/setting-2.svg'
+import Shield from '../../icons/shield-icon.svg'
+import ShieldLock from '../../icons/shield-lock.svg'
+import SidebarIcon from '../../icons/sidebar-icon.svg'
+import SidebarLeft from '../../icons/sidebar-left.svg'
+import SidebarRight from '../../icons/sidebar-right.svg'
+import Signpost from '../../icons/signpost.svg'
+import Slack from '../../icons/slack.svg'
+import Snow from '../../icons/snow-icon.svg'
+import SparksGradientDark from '../../icons/sparks-gradient-dark.svg'
+import SparksGradientLight from '../../icons/sparks-gradient-light.svg'
+import SparksGradient from '../../icons/sparks-gradient.svg'
+import Sparks from '../../icons/sparks.svg'
+import SquareDashed from '../../icons/square-dashed.svg'
+import SshKey from '../../icons/ssh-key.svg'
+import Stack from '../../icons/stack-icon.svg'
+import Star from '../../icons/star-icon.svg'
+import Stop from '../../icons/stop-icon.svg'
+import SubMenuEllipse from '../../icons/sub-menu-ellipse.svg'
+import Success from '../../icons/success.svg'
+import Suggestion from '../../icons/suggestion.svg'
+import SupplyChainGradient from '../../icons/supply-chain-gradient.svg'
+import SupplyChain from '../../icons/supply-chain-icon.svg'
+import Tag from '../../icons/tag.svg'
+import Tasks from '../../icons/tasks.svg'
+import TickCircle from '../../icons/tick-circle.svg'
+import Tick from '../../icons/tick-icon.svg'
+import Ticket from '../../icons/ticket-icon.svg'
+import Trash from '../../icons/trash.svg'
+import TriangleWarning from '../../icons/triangle-warning.svg'
+import TubeSign from '../../icons/tube-sign.svg'
+import Unmerged from '../../icons/unmerged.svg'
+import Unpin from '../../icons/unpin.svg'
+import Upload from '../../icons/upload-icon.svg'
+import User from '../../icons/user-icon.svg'
+import Users from '../../icons/users-icon.svg'
+import Variable from '../../icons/variable-icon.svg'
+import VerticalEllipsis from '../../icons/vertical-ellipsis.svg'
+import WarningTriangleOutline from '../../icons/warning-triangle-outline.svg'
+import Webhook from '../../icons/webhook-icon.svg'
+import Wrench from '../../icons/wrench-icon.svg'
+import XMark from '../../icons/x-mark.svg'
+
+export const IconNameMap = {
+  'chevron-down': ChevronDown,
+  'chevron-right': ChevronRight,
+  'chevron-up': ChevronUp,
+  'filter-list': FilterList,
+  'info-circle': InfoCircle,
+  'double-tick': DoubleTick,
+  play: Play,
+  download: Download,
+  clock: Clock,
+  checkbox: Checkbox,
+  harness: Harness,
+  pipelines: Pipelines,
+  'pipelines-gradient': PipelinesGradient,
+  ellipsis: Ellipsis,
+  repositories: Repositories,
+  'repositories-gradient': RepositoriesGradient,
+  environment: Environment,
+  secrets: Secrets,
+  connectors: Connectors,
+  star: Star,
+  pull: Pull,
+  search: Search,
+  'arrow-long': ArrowLong,
+  success: Success,
+  fail: Fail,
+  'tube-sign': TubeSign,
+  signpost: Signpost,
+  merged: Merged,
+  'pr-draft': Unmerged,
+  comments: Comments,
+  tasks: Tasks,
+  tick: Tick,
+  'no-data-folder': NoDataFolder,
+  'no-data-folder-light': NoDataFolderLight,
+  'no-repository': NoRepository,
+  'no-repository-light': NoRepositoryLight,
+  'no-data-merge': NoDataMerge,
+  'no-data-merge-light': NoDataMergeLight,
+  'no-data-cog': NoDataCog,
+  'no-data-cog-light': NoDataCogLight,
+  'no-data-branches': NoDataBranches,
+  'no-data-branches-light': NoDataBranchesLight,
+  'no-search-magnifying-glass': NoSearchMagnifyingGlass,
+  'no-search-magnifying-glass-light': NoSearchMagnifyingGlassLight,
+  'no-data-members': NoDataMembers,
+  'no-data-members-light': NoDataMembersLight,
+  'no-data-error': NoDataError,
+  'no-data-error-light': NoDataErrorLight,
+  'no-data-pr': NoDataPR,
+  'no-data-pr-light': NoDataPRLight,
+  'no-data-tags': noDataTags,
+  changes: Changes,
+  checks: Checks,
+  'vertical-ellipsis': VerticalEllipsis,
+  clone: Clone,
+  'triangle-warning': TriangleWarning,
+  'pending-clock': PendingClock,
+  circle: Circle,
+  'apple-shortcut': AppleShortcut,
+  'edit-pen': Edit,
+  'gitness-logo': GitnessLogo,
+  'github-actions': GithubActions,
+  folder: Folder,
+  file: File,
+  tag: Tag,
+  branch: Branch,
+  'open-pr': OpenPR,
+  plus: Plus,
+  running: Running,
+  'circle-arrow-top-right': CircleArrowTopRight,
+  'box-lightning': BoxLightning,
+  'box-cloning': BoxCloning,
+  'box-pull-requests': BoxPullRequests,
+  'box-guide': BoxGuide,
+  'node-logo': NodeLogo,
+  'python-logo': PythonLogo,
+  'python-and-node-logo': PythonAndNodeLogo,
+  'ai-sparks': AISparks,
+  sparks: Sparks,
+  'sparks-gradient': SparksGradient,
+  'sparks-gradient-dark': SparksGradientDark,
+  'sparks-gradient-light': SparksGradientLight,
+  'pr-open': PrOpen,
+  'pr-review': PrReview,
+  'pr-closed': PrClosed,
+  'pr-merge': PrMerge,
+  'git-branch': GitBranch,
+  'pr-comment': PrComment,
+  'create-workspace': CreateWorkspace,
+  'create-workspace-light': CreateWorkspaceLight,
+  'no-data-webhooks': NoDataWebhooks,
+  'no-data-webhooks-light': NoDataWebhooksLight,
+  'add-folder': AddFolder,
+  'add-file': AddFile,
+  lightning: Lightning,
+  'harness-plugin': HarnessPlugin,
+  'bitrise-plugin': BitrisePlugin,
+  trash: Trash,
+  'x-mark': XMark,
+  'cog-6': Cog6,
+  pin: Pin,
+  unpin: Unpin,
+  rocket: Rocket,
+  plug: Plug,
+  'shield-lock': ShieldLock,
+  'pr-merged': PRMerged,
+  run: Run,
+  'run-test': RunTest,
+  'ssh-key': SshKey,
+  'green-dot': GreenCircle,
+  'tick-circle': TickCircle,
+  'cancel-grey': CancelGrey,
+  'harness-logo-text': HarnessLogoText,
+  account: Account,
+  'settings-1': Settings1,
+  'settings-2': Settings2,
+  'harness-gradient': HarnessLogoGradient,
+  'harness-gradient-ellipse': HarnessLogoGradientEllipse,
+  'sub-menu-ellipse': SubMenuEllipse,
+  close: Close,
+  notification: Notification,
+  wrench: Wrench,
+  hierarchy: Hierarchy,
+  key: Key,
+  'file-icon': FileIcon,
+  'sidebar-icon': SidebarIcon,
+  'sidebar-left': SidebarLeft,
+  'sidebar-right': SidebarRight,
+  variable: Variable,
+  'clock-icon': ClockIcon,
+  eye: Eye,
+  stack: Stack,
+  'bookmark-icon': BookmarkIcon,
+  webhook: Webhook,
+  user: User,
+  users: Users,
+  'account-icon': AccountIcon,
+  'folder-icon': FolderIcon,
+  briefcase: Briefcase,
+  shield: Shield,
+  snow: Snow,
+  ticket: Ticket,
+  'grid-dots': GridDots,
+  'circle-arrow-top': CircleArrowTop,
+  'circle-arrows-updown': CircleArrowsUpDown,
+  'chevron-fill-down': ChevronFillDown,
+  'more-dots-fill': MoreDotsFill,
+  execution: Execution,
+  'execution-gradient': ExecutionGradient,
+  paint: Paint,
+  navigation: Navigation,
+  logOut: LogOut,
+  'database-gradient': DatabaseGradient,
+  database: Database,
+  'artifacts-gradient': ArtifactsGradient,
+  artifacts: Artifacts,
+  'infrastructure-gradient': InfrastructureGradient,
+  infrastructure: Infrastructure,
+  'flag-gradient': FlagGradient,
+  flag: Flag,
+  'dev-portal-gradient': DevPortalGradient,
+  'dev-portal': DevPortal,
+  'dev-envs-gradient': DevEnvsGradient,
+  'dev-envs': DevEnvs,
+  'dev-insights-gradient': DevInsightsGradient,
+  'dev-insights': DevInsights,
+  'security-tests-gradient': SecurityTestsGradient,
+  'security-tests': SecurityTests,
+  'supply-chain-gradient': SupplyChainGradient,
+  'supply-chain': SupplyChain,
+  'cloud-costs-gradient': CloudCostsGradient,
+  'cloud-costs': CloudCosts,
+  'incidents-gradient': IncidentsGradient,
+  incidents: Incidents,
+  'chaos-engineering-gradient': ChaosEngineeringGradient,
+  'chaos-engineering': ChaosEngineering,
+  'dashboards-gradient': DashboardsGradient,
+  dashboards: Dashboards,
+  'menu-dots': MenuDots,
+  'bookmark-add': BookmarkAdd,
+  upload: Upload,
+  minus: Minus,
+  header: Header,
+  bold: Bold,
+  attachment: Attachment,
+  code: Code,
+  list: List,
+  checklist: Checklist,
+  italicize: Italicize,
+  'no-data-commits': NoDataCommits,
+  'no-data-commits-light': NoDataCommitsLight,
+  compare: Compare,
+  lock: Lock,
+  cross: Cross,
+  'collapse-diff': CollapseDiff,
+  'expand-diff': ExpandDiff,
+  'circle-plus': CirclePlus,
+  'code-brackets': CodeBrackets,
+  'attachment-image': AttachmentImage,
+  'collapse-comment': CollapseComment,
+  'expand-comment': ExpandComment,
+  suggestion: Suggestion,
+  'repo-icon': RepoIcon,
+  arrow: Arrow,
+  'arrow-short': ArrowShort,
+  'bold-plus': BoldPlus,
+  'checkbox-circle': CheckboxCircle,
+  'circle-with-sector': CircleWithSector,
+  'command-symbol': CommandSymbol,
+  'cross-circle': CrossCircle,
+  'warning-triangle-outline': WarningTriangleOutline,
+  slack: Slack,
+  docker: Docker,
+  stop: Stop,
+  'collapse-out': CollapseOut,
+  'collapse-in': CollapseIn,
+  'square-dashed': SquareDashed
+} satisfies Record<string, React.FunctionComponent<React.SVGProps<SVGSVGElement>>>
