+++ packages/ui/src/components/icon/index.ts
@@ -1,0 +1,2 @@
+export * from './icon-name-map'
+export * from './icon'
