+++ packages/ui/src/components/index.ts
@@ -18,7 +18,6 @@ export * from './avatar'
 export * from './navbar'
 export * from './navbar-skeleton'
 export * from './theme-selector'
-export * from './theme-selector/types'
 export * from './theme-selector/utils'
 export * from './no-data'
 export * from './tabs'
