+++ packages/ui/src/components/navbar/navbar-user/index.tsx
@@ -1,18 +1,7 @@
 import { useMemo } from 'react'
 
-import {
-  Avatar,
-  ColorType,
-  ContrastType,
-  DropdownMenu,
-  FullTheme,
-  getModeColorContrastFromFullTheme,
-  Icon,
-  IThemeStore,
-  ModeType,
-  Text
-} from '@/components'
-import { useRouterContext } from '@/context'
+import { Avatar, DropdownMenu, getModeColorContrastFromFullTheme, Icon, Text } from '@/components'
+import { ColorType, ContrastType, FullTheme, ModeType, useRouterContext, useTheme } from '@/context'
 import { TypesUser } from '@/types'
 import { cn } from '@utils/cn'
 import { getInitials } from '@utils/stringUtils'
@@ -57,20 +46,13 @@ interface NavbarUserProps {
   currentUser: TypesUser | undefined
   handleCustomNav: () => void
   handleLogOut: () => void
-  useThemeStore: () => IThemeStore
   useTranslationStore: () => TranslationStore
 }
 
-export const NavbarUser = ({
-  currentUser,
-  handleCustomNav,
-  handleLogOut,
-  useThemeStore,
-  useTranslationStore
-}: NavbarUserProps) => {
+export const NavbarUser = ({ currentUser, handleCustomNav, handleLogOut, useTranslationStore }: NavbarUserProps) => {
   const { Link } = useRouterContext()
   const username = currentUser?.display_name || currentUser?.uid || ''
-  const { theme, setTheme } = useThemeStore()
+  const { theme, setTheme } = useTheme()
   const { t, i18n, changeLanguage } = useTranslationStore()
   const { mode, color, contrast } = useMemo(() => getModeColorContrastFromFullTheme(theme), [theme])
