+++ packages/ui/src/context/theme/index.ts
@@ -1,0 +1,2 @@
+export * from './types'
+export * from './theme-context'
