+++ packages/ui/src/utils/is-light-theme.ts
@@ -1,0 +1,5 @@
+import { defaultTheme, FullTheme, ModeType } from '@/context'
+
+export const isLightTheme = (theme?: FullTheme) => {
+  return (theme ?? defaultTheme).includes(ModeType.Light)
+}
