+++ packages/ui/src/views/repo/pull-request/pull-request-list-page.tsx
@@ -23,7 +23,7 @@ type PRListFiltersKeys = keyof PRListFilters
 
 const PRListFilterHandler = createFilters<PRListFilters>()
 
-const PullRequestList: FC<PullRequestPageProps> = ({
+const PullRequestListPage: FC<PullRequestPageProps> = ({
   usePullRequestListStore,
   spaceId,
   repoId,
@@ -314,4 +314,4 @@ const PullRequestList: FC<PullRequestPageProps> = ({
     </SandboxLayout.Main>
   )
 }
-export { PullRequestList }
+export { PullRequestListPage }
